Load small RNA-seq bcbio-nextgen run

Simply point to the final upload directory output by bcbio-nextgen, and this function will take care of the rest. It automatically imports small RNA-seq counts, metadata, and program versions used.

loadSmallRnaRun(projectDir = "date-final",
  interestingGroups = "description", maxSamples = 50, dataDir = NULL,
  colData = NULL, ...)

Arguments

projectDir
Path to final upload directory. This path is set when running bcbio_nextgen -w template.

interestingGroups
Character vector of interesting groups. First entry is used for plot colors during quality control (QC) analysis. Entire vector is used for PCA and heatmap QC functions.

maxSamples
Optional. Maximum number of samples to calculate rlog and variance stabilization object from DESeq2.

dataDir
Folder to keep a cache of the object.

colData
Optional External metadata to be used while reading samples.

...
Additional arguments, saved as metadata.

Value

bcbioSmallRnaDataSet.

Note

When working in RStudio, we recommend connecting to the bcbio-nextgen run directory as a remote connection over sshfs.

Examples

path <- system.file("extra", package="bcbioSmallRna")
bcb <- loadSmallRnaRun(file.path(path, "bcbio"), "country")

#> /Users/lpantano/repos/hbc/bcbioSmallRna/inst/extra/bcbio

#> Error in loadSmallRnaRun(file.path(path, "bcbio"), "country"): YAML project summary missing /Users/lpantano/repos/hbc/bcbioSmallRna/inst/extra/bcbio/project-summary.yaml